refactor: change appearance type

+++ starwars/src/main/java/online/cotefacil/starwars/model/Planet.java
@@ -20,5 +20,5 @@ public class Planet {
     private String name;
     private String climate;
     private String terrain;
-    private Long appearance;
+    private int appearance;
 }
